Split AssessmentServiceTest class up into small chunks
